+++ directive/ref.js
@@ -1,4 +1,4 @@
-import { directive } from "../src/core.js";
+import { directive } from "../core.js";
 
 // ref must be last within primaries, since that must be skipped by :each, but before secondaries
 directive.ref = (el, expr, state) => {
