+++ signal.js
@@ -1,0 +1,6 @@
+import sprae from './core.js'
+import * as signals from 'ulive'
+
+Object.assign(sprae, signals)
+
+export * from 'ulive';
